Cleanup
Deleted a couple of unused files and moved other files to utils
