+++ src/main/cpp/Robot.h
@@ -27,7 +27,6 @@ class DragonField;
 class AutonPreviewer;
 class RobotState;
 class SomeMech;
-class RobotDefinition;
 class DragonDataLoggerMgr;
 
 class Robot : public frc::TimedRobot
@@ -69,7 +68,6 @@ private:
     AutonPreviewer *m_previewer;
     RobotState *m_robotState;
     SomeMech *m_someMech;
-    RobotDefinition *m_robot;
     DragonDataLoggerMgr *m_datalogger;
     bool isFMSAttached = false;
 };
