+++ src/main/cpp/utils/logging/DragonDataLogger.cpp
@@ -31,69 +31,6 @@ void DragonDataLogger::LogBoolData(DragonDataLoggerSignals::BoolSignals signalID
     auto signals = DragonDataLoggerSignals::GetInstance();
     if (signals != nullptr)
     {
-        /**
-        switch (signalID)
-        {
-        case DragonDataLoggerSignals::BoolSignals::NOTE_MANAGER_HAS_VISION:
-            if (value != signals->m_currHasVision)
-            {
-                signals->m_hasVision.Append(value);
-                signals->m_currHasVision = value;
-            }
-            break;
-        case DragonDataLoggerSignals::BoolSignals::NOTE_MANAGER_BACK_SENSOR:
-            if (value != signals->m_currBackSensor)
-            {
-                signals->m_backSensor.Append(value);
-                signals->m_currBackSensor = value;
-            }
-            break;
-        case DragonDataLoggerSignals::BoolSignals::NOTE_MANAGER_FEEDER_SENSOR:
-            if (value != signals->m_currFeederSensor)
-            {
-                signals->m_feederSensor.Append(value);
-                signals->m_currFeederSensor = value;
-            }
-            break;
-        case DragonDataLoggerSignals::BoolSignals::NOTE_MANAGER_FRONT_SENSOR:
-            if (value != signals->m_currFrontSensor)
-            {
-                signals->m_frontSensor.Append(value);
-                signals->m_currFrontSensor = value;
-            }
-            break;
-        case DragonDataLoggerSignals::BoolSignals::NOTE_MANAGER_LAUNCHER_SENSOR:
-            if (value != signals->m_currLauncherSensor)
-            {
-                signals->m_launcherSensor.Append(value);
-                signals->m_currLauncherSensor = value;
-            }
-            break;
-        case DragonDataLoggerSignals::BoolSignals::NOTE_MANAGER_PLACER_IN_SENSOR:
-            if (value != signals->m_currPlacerInSensor)
-            {
-                signals->m_placerInSensor.Append(value);
-                signals->m_currPlacerInSensor = value;
-            }
-            break;
-        case DragonDataLoggerSignals::BoolSignals::NOTE_MANAGER_PLACER_MID_SENSOR:
-            if (value != signals->m_currPlacerMidSensor)
-            {
-                signals->m_placerMidSensor.Append(value);
-                signals->m_currPlacerMidSensor = value;
-            }
-            break;
-        case DragonDataLoggerSignals::BoolSignals::NOTE_MANAGER_PLACER_OUT_SENSOR:
-            if (value != signals->m_currPlacerOutSensor)
-            {
-                signals->m_placerOutSensor.Append(value);
-                signals->m_currPlacerOutSensor = value;
-            }
-            break;
-        default:
-            break;
-        }
-        **/
     }
 }
 
@@ -118,63 +55,7 @@ void DragonDataLogger::LogDoubleData(DragonDataLoggerSignals::DoubleSignals sign
                 signals->m_currChassisYaw = value;
             }
             break;
-            /**
-        case DragonDataLoggerSignals::DoubleSignals::NOTE_MANAGER_TARGET_ANGLE_DEGREES:
-            if (std::abs(value - signals->m_currNmTargetAngle) > m_doubleTolerance)
-            {
-                signals->m_nmTargetAngle.Append(value);
-                signals->m_currNmTargetAngle = value;
-            }
-            break;
 
-        case DragonDataLoggerSignals::DoubleSignals::NOTE_MANAGER_ACTUAL_ANGLE_DEGREES:
-            if (std::abs(value - signals->m_currNmActualAngle) > m_doubleTolerance)
-            {
-                signals->m_nmActualAngle.Append(value);
-                signals->m_currNmActualAngle = value;
-            }
-            break;
-
-        case DragonDataLoggerSignals::DoubleSignals::NOTE_MANAGER_TARGET_TOP_WHEEL_SPEED_RPM:
-            if (std::abs(value - signals->m_currNmTopTarget) > m_doubleTolerance)
-            {
-                signals->m_nmTopTarget.Append(value);
-                signals->m_currNmTopTarget = value;
-            }
-            break;
-
-        case DragonDataLoggerSignals::DoubleSignals::NOTE_MANAGER_TARGET_BOTTOM_WHEEL_SPEED_RPM:
-            if (std::abs(value - signals->m_currNmBottomTarget) > m_doubleTolerance)
-            {
-                signals->m_nmBottomTarget.Append(value);
-                signals->m_currNmBottomTarget = value;
-            }
-            break;
-
-        case DragonDataLoggerSignals::DoubleSignals::NOTE_MANAGER_ACTUAL_TOP_WHEEL_SPEED_RPM:
-            if (std::abs(value - signals->m_currNmTopActual) > m_doubleTolerance)
-            {
-                signals->m_nmTopActual.Append(value);
-                signals->m_currNmTopActual = value;
-            }
-            break;
-
-        case DragonDataLoggerSignals::DoubleSignals::NOTE_MANAGER_ACTUAL_BOTTOM_WHEEL_SPEED_RPM:
-            if (std::abs(value - signals->m_currNmBottomActual) > m_doubleTolerance)
-            {
-                signals->m_nmBottomActual.Append(value);
-                signals->m_currNmBottomActual = value;
-            }
-            break;
-
-        case DragonDataLoggerSignals::DoubleSignals::NOTE_MANAGER_DISTANCE_FROM_SPEAKER_METERS:
-            if (std::abs(value - signals->m_currDistFromSpeaker) > m_doubleTolerance)
-            {
-                signals->m_distFromSpeaker.Append(value);
-                signals->m_currDistFromSpeaker = value;
-            }
-            break;
-        **/
         default:
             break;
         }
@@ -203,16 +84,6 @@ void DragonDataLogger::LogStringData(DragonDataLoggerSignals::StringSignals sign
                 signals->m_currHeadingState = value;
             }
             break;
-
-            /**
-            case DragonDataLoggerSignals::StringSignals::NOTE_MANAGER_STATE:
-                if (signals->m_currNoteMgrState != value)
-                {
-                    signals->m_noteMgrState.Append(value);
-                    signals->m_currNoteMgrState = value;
-                }
-                break;
-                **/
 
         default:
             break;
